+++ packages/profile-sync-controller/src/controllers/user-storage/UserStorageController.ts
@@ -186,10 +186,6 @@ type ActionsObj = CreateActionsObj<
   | 'performDeleteStorage'
   | 'performBatchDeleteStorage'
   | 'getStorageKey'
-  | 'enableProfileSyncing'
-  | 'disableProfileSyncing'
-  | 'syncInternalAccountsWithUserStorage'
-  | 'saveInternalAccountToUserStorage'
 >;
 export type UserStorageControllerGetStateAction = ControllerGetStateAction<
   typeof controllerName,
@@ -211,14 +207,6 @@ export type UserStorageControllerPerformDeleteStorage =
 export type UserStorageControllerPerformBatchDeleteStorage =
   ActionsObj['performBatchDeleteStorage'];
 export type UserStorageControllerGetStorageKey = ActionsObj['getStorageKey'];
-export type UserStorageControllerEnableProfileSyncing =
-  ActionsObj['enableProfileSyncing'];
-export type UserStorageControllerDisableProfileSyncing =
-  ActionsObj['disableProfileSyncing'];
-export type UserStorageControllerSyncInternalAccountsWithUserStorage =
-  ActionsObj['syncInternalAccountsWithUserStorage'];
-export type UserStorageControllerSaveInternalAccountToUserStorage =
-  ActionsObj['saveInternalAccountToUserStorage'];
 
 export type AllowedActions =
   // Keyring Requests
@@ -442,26 +430,6 @@ export default class UserStorageController extends BaseController<
     this.messagingSystem.registerActionHandler(
       'UserStorageController:getStorageKey',
       this.getStorageKey.bind(this),
-    );
-
-    this.messagingSystem.registerActionHandler(
-      'UserStorageController:enableProfileSyncing',
-      this.enableProfileSyncing.bind(this),
-    );
-
-    this.messagingSystem.registerActionHandler(
-      'UserStorageController:disableProfileSyncing',
-      this.disableProfileSyncing.bind(this),
-    );
-
-    this.messagingSystem.registerActionHandler(
-      'UserStorageController:syncInternalAccountsWithUserStorage',
-      this.syncInternalAccountsWithUserStorage.bind(this),
-    );
-
-    this.messagingSystem.registerActionHandler(
-      'UserStorageController:saveInternalAccountToUserStorage',
-      this.saveInternalAccountToUserStorage.bind(this),
     );
   }